chore(refactor): Extract class refactoring
