+++ src/main/java/io/jenkins/plugins/opentelemetry/OtelConstantUtils.java
@@ -1,0 +1,26 @@
+package io.jenkins.plugins.opentelemetry;
+
+public class OtelConstantUtils {
+
+    public static final String FREESTYLE = "freestyle";
+
+    public static final String MATRIX = "matrix";
+
+    public static final String MAVEN = "maven";
+
+    public static final String MULTIBRANCH = "multibranch";
+
+    public static final String WORKFLOW = "workflow";
+
+    public static final String UNKNOWN = "unknown";
+
+    public static final String BRANCH = "branch";
+
+    public static final String CHANGE_REQUEST = "change_request";
+
+    public static final String TAG = "tag";
+
+    public static final String JENKINS_CORE = "jenkins-core";
+
+    public static final String UNKNOWN_VALUE = "#unknown";
+}
