+++ src/main/java/io/jenkins/plugins/opentelemetry/OtelUtils.java
@@ -51,20 +51,9 @@ import java.util.Optional;
 import java.util.TreeMap;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+import static io.jenkins.plugins.opentelemetry.OtelConstantUtils.*;
 
 public class OtelUtils {
-
-    public static final String FREESTYLE = "freestyle";
-    public static final String MATRIX = "matrix";
-    public static final String MAVEN = "maven";
-    public static final String MULTIBRANCH = "multibranch";
-    public static final String WORKFLOW = "workflow";
-    public static final String UNKNOWN = "unknown";
-    public static final String BRANCH = "branch";
-    public static final String CHANGE_REQUEST = "change_request";
-    public static final String TAG = "tag";
-    public static final String JENKINS_CORE = "jenkins-core";
-    public static final String UNKNOWN_VALUE = "#unknown";
 
     @CheckForNull
     public static String getSystemPropertyOrEnvironmentVariable(String environmentVariableName) {
